requirements.txt
Django~=2.2.4
Djangorestframework~=3.10.2
django-cors-headers~=3.0.2